Move service init code to nav base activity
