+++ app/src/main/java/io/islnd/android/islnd/app/activities/NavBaseActivity.java
@@ -1,6 +1,8 @@
 package io.islnd.android.islnd.app.activities;
 
 import android.Manifest;
+import android.accounts.Account;
+import android.accounts.AccountManager;
 import android.app.AlertDialog;
 import android.content.ContentResolver;
 import android.content.Context;
@@ -34,6 +36,7 @@ import android.widget.TextView;
 import com.google.zxing.integration.android.IntentIntegrator;
 import com.google.zxing.integration.android.IntentResult;
 import io.islnd.android.islnd.app.database.DataUtils;
+import io.islnd.android.islnd.app.database.IslndContract;
 import io.islnd.android.islnd.app.database.IslndDb;
 import io.islnd.android.islnd.app.R;
 import io.islnd.android.islnd.app.fragments.FeedFragment;
@@ -58,6 +61,7 @@ public class NavBaseActivity extends AppCompatActivity
     private EditText mSmsEditText = null;
     private View mDialogView = null;
     private Context mContext;
+    private ContentResolver mResolver;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -70,6 +74,15 @@ public class NavBaseActivity extends AppCompatActivity
         Fragment fragment = new FeedFragment();
         FragmentManager fragmentManager = getSupportFragmentManager();
         fragmentManager.beginTransaction().replace(R.id.content_frame, fragment).commit();
+
+        // Create sync account and force sync
+        Account account = createSyncAccount(this);
+        mResolver = getContentResolver();
+        mResolver.setSyncAutomatically(
+                account,
+                getString(R.string.content_authority),
+                true);
+        mResolver.requestSync(account, IslndContract.CONTENT_AUTHORITY, new Bundle());
     }
 
     private void onCreateDrawer() {
@@ -266,7 +279,8 @@ public class NavBaseActivity extends AppCompatActivity
 
         mSmsEditText = (EditText) mDialogView.findViewById(R.id.sms_number_edit_text);
 
-        builder.setPositiveButton(getString(R.string.send), (DialogInterface dialog, int id) -> {
+        builder.setPositiveButton(
+                getString(R.string.send), (DialogInterface dialog, int id) -> {
                     sendSms(mSmsEditText.getText().toString());
                 })
                 .setNegativeButton(android.R.string.cancel, null)
@@ -346,7 +360,7 @@ public class NavBaseActivity extends AppCompatActivity
         String contactNumber = null;
 
         // getting contacts ID
-        ContentResolver cr = getContentResolver();
+        ContentResolver cr = mResolver;
         Cursor cursorID = cr.query(uriContact,
                 new String[]{ContactsContract.Contacts._ID},
                 null, null, null);
@@ -355,7 +369,7 @@ public class NavBaseActivity extends AppCompatActivity
             String contactId =
                     cursorID.getString(cursorID.getColumnIndex(ContactsContract.Contacts._ID));
 
-            Cursor cursorPhone = getContentResolver().query(
+            Cursor cursorPhone = mResolver.query(
                     ContactsContract.CommonDataKinds.Phone.CONTENT_URI,
                     new String[]{ContactsContract.CommonDataKinds.Phone.NUMBER},
 
@@ -408,5 +422,16 @@ public class NavBaseActivity extends AppCompatActivity
                 })
                 .setNegativeButton(android.R.string.cancel, null)
                 .show();
+    }
+
+    public static Account createSyncAccount(Context context) {
+        Account newAccount = new Account(
+                context.getString(R.string.sync_account),
+                context.getString(R.string.sync_account_type));
+
+        AccountManager accountManager = (AccountManager) context.getSystemService(context.ACCOUNT_SERVICE);
+        accountManager.addAccountExplicitly(newAccount, null, null);
+
+        return newAccount;
     }
 }
